오류 처리 및 리스폰스 타입 변경

+++ domain/src/main/java/com/a90ms/domain/data/dto/game/ChampionsDto.kt
@@ -9,8 +9,8 @@ data class ChampionsDto(
     val kills: Int,
     val deaths: Int,
     val assists: Int,
-    val wins: Int,
-    val losses: Int,
+    val wins: Float,
+    val losses: Float,
     val winPerRate: Int,
     val winPerRateToString: String
 )
