+++ domain/src/main/java/com/a90ms/domain/data/dto/summoner/LeagueDto.kt
@@ -2,8 +2,8 @@ package com.a90ms.domain.data.dto.summoner
 
 data class LeagueDto(
     val hasResults: Boolean,
-    val wins: Int,
-    val losses: Int,
+    val wins: Float,
+    val losses: Float,
     val tierRank: TierRankDto,
     val winningRate: String
 )
